% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_alphas.R
\name{compare_alphas}
\alias{compare_alphas}
\title{Graphical display to investigate the impact of concentration parameter values in the Dirichlet distribution.}
\usage{
compare_alphas(n, alpha)
}
\arguments{
\item{n}{number of weights which is generated from Dirichlet distribution}

\item{alpha}{concentration parameter is given in vector form; each element represents each group.}
}
\value{
graphical display of different concentration parameters of Dirichlet distribution
}
\description{
This function provides a graphical display to investigate the impact of concentration parameter values in the Dirichlet distribution.
}
\details{
We assumed that the concentration parameters are identical in all weights, it means that \eqn{\boldsymbol{\alpha} =[\alpha, \alpha, \cdots \alpha]}. The probability mass
function of the weights is given by: \deqn{\displaystyle{f(w_1,w_2, \cdots w_n;\boldsymbol{\alpha} )=\frac{\displaystyle{\Gamma(n\alpha)}} {{\Gamma(\alpha)}^n} \prod_{i=1}^{n}{w_i^{\alpha-1}}}}
where \eqn{\boldsymbol{\alpha} =[\alpha, \alpha, \cdots \alpha]} is the vector of concentration parameter, and \eqn{w_i} is the stochastic weights which sum to one.
}
\examples{
n <- 10000
alpha <- c(5,10,20)
compare_alphas(n,alpha)
}
