% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mixing.R
\name{compare_mixing}
\alias{compare_mixing}
\title{The graphical comparison between different mixing schemes with varying parameters of mixing by simulation results.}
\usage{
compare_mixing(n_iter, mu, sigma, b, k, distribution)
}
\arguments{
\item{n_iter}{number of iterations}

\item{mu}{average number of colony-forming units in a primary sample which is in logarithmic scale if we use a lognormal distribution}

\item{sigma}{log standard deviation of the colony-forming units in a primary sample}

\item{b}{concentration parameter}

\item{k}{number of small portions/ primary samples}

\item{distribution}{what suitable distribution type we have employed for simulation such as 'Poisson-Type A' or 'Poisson-Type B' or 'Lognormal-Type A' or 'Lognormal-Type B'}
}
\value{
graphical comparison between different mixing schemes
}
\description{
This function provides the graphical displays for a different set of mixing plans for comparison purpose. (to be finished later on)
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by mixing of \eqn{k} primary sample and \eqn{N' = \sum(N_i)}  and \eqn{i} be the number of colony-forming units in the \eqn{i^{th}} primary sample; where \eqn{i = 1,2,....k} and \eqn{y_i=x_i/\sum(x_i) = q_i/Q}; where \eqn{x_i} follows \eqn{gamma (b,1)} (to be finished later on)
\itemize{
\item Case 1 (Poisson-Type A): \eqn{N_i} follows \eqn{Poisson(\mu/k)}
\item Case 2 (Poisson-Type B): \eqn{N_i} follows \eqn{Poisson(\mu*y_i)}
\item Case 3 (Lognormal-Type A): \eqn{N_i} follows \eqn{Binomial(M_i,1/k)}; where  \eqn{M_i} follows \eqn{Lognormal(\mu, \sigma)}
\item Case 4 (Lognormal-Type B): \eqn{N_i}  follows \eqn{Binomial(M_i,y_i)}; where  \eqn{M_i} follows \eqn{Lognormal(\mu, \sigma)}
}
}
\examples{
n_iter <- 2000000
mu <- c(log(100),log(100),log(100))
sigma <- c(0.8,0.8,0.8)
b <- c(0.1,1,10)
k <- c(30,30,30)
distribution <-  c("Lognormal-Type B","Lognormal-Type B","Lognormal-Type B")
compare_mixing (n_iter, mu, sigma, b, k, distribution )
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_multiple}, \link{sim_single}
}
