% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mixing_1.R
\name{compare_mixing_1}
\alias{compare_mixing_1}
\title{Graphical comparison of mixing plans based on the estimated probability of detection at each revolution.}
\usage{
compare_mixing_1(mu, sigma, alpha_in, k, l, rate, distribution, UDL, n_sim)
}
\arguments{
\item{mu}{the average number of CFUs (\eqn{\mu}) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units (CFUs) in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{rate}{concentration parameter changing rate in each of the revolutions}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{UDL}{the upper decision limit, which depends on the type of microorganisms and testing regulations.}

\item{n_sim}{number of simulations}
}
\value{
graphical display of estimated probability of detection at each revolution in the mixing.
}
\description{
This function provides a graphical display to compare mixing plans based on the estimated probability of detection at each revolution of the mixing process.
}
\examples{
mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- c(30,50,75)
l <- 25000
rate <- 0.01
distribution <-  "Poisson lognormal-Type B"
UDL <- 0
n_sim <- 2000
compare_mixing_1(mu,sigma , alpha_in, k, l, rate, distribution, UDL, n_sim)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_single_pd_stages}
}
