% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mixing_2.R
\name{compare_mixing_2}
\alias{compare_mixing_2}
\title{Graphical comparison of mixing plans based on estimated probability of detection at the end of the mixing with different revolutions}
\usage{
compare_mixing_2(
  mulower,
  muupper,
  sigma,
  alpha_in,
  k,
  l,
  r,
  distribution,
  UDL,
  n_sim
)
}
\arguments{
\item{mulower}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muupper}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sigma}{the standard deviation of the colony-forming units (CFUs) in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{r}{the rate of the concentration parameter changes at each mixing stage}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{UDL}{the upper decision limit, which depends on the type of microorganisms and testing regulations.}

\item{n_sim}{number of simulations}
}
\value{
graphical display compares mixing plans using a different number of revolutions based on the estimated probability of detection at the end of the mixing.
}
\description{
This function provides a graphical display to compare mixing plans using a different number of revolutions based on the estimated probability of detection at the end of mixing.
}
\examples{
mulower <- 0.1
muupper <- 200
sigma <- 0.8
alpha_in <- 0.01
k <- 30
l <- c(500,5000)
r <- 0.01
distribution <-  "Poisson lognormal-Type B"
UDL <- 0
n_sim <- 2000
compare_mixing_2(mulower, muupper, sigma , alpha_in, k, l, r, distribution, UDL, n_sim)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_single_pd_stages}
}
