% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_usl.R
\name{mixing_usl}
\alias{mixing_usl}
\title{Graphical display to find out upper stabilizing limit (USL).}
\usage{
mixing_usl(mu, sigma, alpha_in, k, l, rate, distribution, n_sim)
}
\arguments{
\item{mu}{the average number of CFUs (\eqn{\mu}) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{rate}{concentration parameter changing rate in the each revolutions}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{n_sim}{number of simulations}
}
\value{
graphical display to find out upper stabilizing limit (USL).
}
\description{
This function provides a graphical display to find out upper stabilizing limit (USL). Based on the cumulative mean at each stage of the mixing,
the cumulative mean persists in the same value after a particular number of stages. Therefore, the upper stabilizing limit (USL) is defined as
a stabilizing point in the graphical display of the expected total number of CFUs versus the number of revolutions when the cumulative mean remains stable.
}
\examples{
mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- 30
l <- 250
rate <- 0.01
distribution <-  "Poisson lognormal-Type B"
n_sim <- 20
mixing_usl(mu, sigma, alpha_in, k, l, rate, distribution, n_sim)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_single_stages}
}
