% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multiple.R
\name{sim_multiple}
\alias{sim_multiple}
\title{The total number of colony-forming units in the mixed sample by simulation result (in the multiple mixing plan).}
\usage{
sim_multiple(n_iter, mu, sigma, b, k, distribution)
}
\arguments{
\item{n_iter}{number of iterations}

\item{mu}{average number of colony-forming units in a primary sample which is in logarithmic scale if we use a lognormal distribution}

\item{sigma}{log standard deviation of the colony-forming units in a primary sample}

\item{b}{concentration parameter}

\item{k}{number of small portions/ primary samples}

\item{distribution}{what suitable distribution type we have employed for simulation such as 'Poisson-fair' or 'Poisson-beta' or 'Lognormal-fair' or 'Lognormal-beta'}
}
\value{
total number of colony forming units in the multiple mixing scheme
}
\description{
This function calculates the resulting total number of colony forming units in the mixed sample in the multiple mixing plans. (to be finished later on)
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by mixing of \eqn{k} primary sample and \eqn{N' = \sum(N_i)} (to be finished later on)
}
\examples{
n_iter <- 200000
mu <- c(20,30,50,60)
sigma <- c(0.8,0.8,0.8,0.8)
b <- c(0.1,0.1,0.1,0.1)
k <- c(10,10,10,10)
distribution <-  c("Poisson-fair","Poisson-fair","Poisson-fair","Poisson-fair")
head(sim_multiple( n_iter, mu, sigma, b, k, distribution))
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_single}, \link{compare_mixing}
}
