% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multiple_prevalence.R
\name{sim_multiple_prevalence}
\alias{sim_multiple_prevalence}
\title{The estimated average prevalence value in the multiple mixing plans.}
\usage{
sim_multiple_prevalence(
  mu,
  sigma,
  alpha_in,
  k,
  l,
  rate,
  distribution,
  UL,
  n_sim
)
}
\arguments{
\item{mu}{the average number of colony-forming units (\eqn{\mu}) in the mixed sample, which is in logarithmic scale if we use a Lognormal / Poisson lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{rate}{concentration parameter changing rate in the each revolutions}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{UL}{the upper limit value of the expected total CFU, which can be found from stabilising point when the mean is about standard deviation}

\item{n_sim}{number of simulations}
}
\value{
the average prevalence in the mixing process after a specific number of revolutions.
}
\description{
This function estimates the average prevalence in the mixing process after a specific number of revolutions for different mixing schemes.
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by contribution of \eqn{k} primary samples mixing, \eqn{N' = \sum N_i} and \eqn{l} be the number of stages in the mixing process. This function estimates the average prevalence value after a specific number of revolutions in each mixing scheme. However, we need to apply function \link{sim_single_prevalence} if we want to estimate individual prevalence values at each stage of the mixing process.
}
\examples{
mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- c(30,75)
l <- 25000
rate <- 0.01
distribution <- c("Poisson lognormal-Type B","Poisson lognormal-Type B")
UL <- 138
n_sim <- 20000
sim_multiple_prevalence(mu, sigma , alpha_in, k, l, rate, distribution, UL, n_sim)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_single_prevalence}
}
