% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pois_stages.R
\name{sim_pois_stages}
\alias{sim_pois_stages}
\title{A graphical display of the mean and variance changes at each mixing stage.}
\usage{
sim_pois_stages(mu, sigma, alpha_in, k, l, rate, distribution, n_sim)
}
\arguments{
\item{mu}{the average number of CFUs (\eqn{\mu}) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{rate}{concentration parameter changing rate in each of the revolutions}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{n_sim}{number of simulations}
}
\value{
Mean and variance changes at each mixing stage.
}
\description{
This function creates a graphical display of the mean and variance changes at each mixing stage.
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by mixing of \eqn{k} primary samples and \eqn{N' = \sum N_i}.
This function produces a graphical display of the mean and variance changes at each mixing stage. It is helpful to identify the optimal number of revolutions of the mixture, which is a point of mixing that initiates Poisson-like homogeneity.
}
\examples{
mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- 30
l <- 5000
rate <- 0.001
distribution <-  "Poisson lognormal-Type B"
n_sim <- 20000
sim_pois_stages(mu, sigma , alpha_in, k, l, rate, distribution, n_sim)
}
