% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single_detectability.R
\name{sim_single_detectability}
\alias{sim_single_detectability}
\title{The estimated detectability in the single mixing scheme.}
\usage{
sim_single_detectability(
  mu,
  sigma,
  alpha_in,
  k,
  l,
  rate,
  distribution,
  UDL,
  n_sim
)
}
\arguments{
\item{mu}{the average number of CFUs (\eqn{\mu}) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{rate}{concentration parameter changing rate in the each revolutions}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{UDL}{the upper decision limit of the expected total CFUs, which can be found from stabilising point when the mean is about standard deviation}

\item{n_sim}{number of simulations}
}
\value{
the estimated detectability at each stage of the mixing process.
}
\description{
This function provides an estimated detectability value in the mixing process of the single mixing scheme.
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by contribution of \eqn{k} primary samples mixing, \eqn{N' = \sum N_i} and \eqn{l} be the number of stages in the mixing process.
This function provides an estimated detectability in the mixing process of the single mixing scheme. To more details, please refer the details section of  \link{compare_mixing_detectability}.
}
\examples{
mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- 30
l <- 25000
rate <- 0.01
distribution <-  "Poisson lognormal-Type B"
UDL <- 138
n_sim <- 20000
sim_single_detectability(mu, sigma , alpha_in, k, l, rate, distribution, UDL, n_sim)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
\item McCallum DA (2005) A conceptual guide to detection probability for point counts and other count-based survey methods. USDA Forest Service General Technical Report PSW-GTR-191, \href{https://www.fs.usda.gov/treesearch/pubs/32063}{pp 754-761}.
}
}
\seealso{
\link{sim_single}
}
