% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multiple_stages.R
\name{sim_multiple_stages}
\alias{sim_multiple_stages}
\title{The expected total number of CFUs in the mixed sample in the multiple mixing schemes at each stage of the mixing process.}
\usage{
sim_multiple_stages(mu, sigma, alpha_in, k, l, rate, distribution, n_sim)
}
\arguments{
\item{mu}{the average number of CFUs (\eqn{\mu}) in the mixed sample, which is in a logarithmic scale if we use a Lognormal / Poisson lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units in the mixed sample on the logarithmic scale (default value 0.8)}

\item{alpha_in}{concentration parameter at the initial stage}

\item{k}{number of small portions / primary samples}

\item{l}{number of revolutions / stages}

\item{rate}{concentration parameter changing rate in the each revolutions}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"} or \code{"Poisson lognormal-Type A"} or \code{"Poisson lognormal-Type B"}}

\item{n_sim}{number of simulations}
}
\value{
The expected total number of CFUs in each revolution / stage.
}
\description{
This function calculates the resulting expected total number of CFUs in the mixed sample in the multiple mixing plans at each stage of the mixing process.
}
\details{
Let \eqn{N'} be the number of CFUs in the mixed sample, which is produced by the mixing of \eqn{k}  primary samples and \eqn{N' = \sum N_i} and let \eqn{N_i} be the number of CFUs.
For this package development, we have employed the notations 'Type-A' and 'Type-B' to indicate the type of distributions, which are applied in the previous literature as 'fair' and 'beta', respectively; see \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{Nauta (2005)}.

This package will consider stage-by-stage the mixing process and assumes systematically breaking clusters at every stage of the mixing. Therefore, it can be assumed the concentration parameter also systematically changes with the concentration of the contribution.
}
\examples{
rate <- 0.01
l <- 25000
mu <- 100
sigma <- 0.8
alpha_in <- 0.01
k <- c(30,60)
distribution <- c("Poisson lognormal-Type B","Poisson lognormal-Type B")
n_sim <- 2000
colMeans(sim_multiple_stages(mu, sigma, alpha_in, k, l, rate, distribution, n_sim))
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_single}
}
