% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single.R
\name{sim_single}
\alias{sim_single}
\title{The total number of colony-forming units in the mixed sample by simulation result (in the single mixing plan).}
\usage{
sim_single(mu, sigma, alpha, k, distribution, n_sim, summary = FALSE)
}
\arguments{
\item{mu}{the average number of colony-forming units in the mixed sample, which is in logarithmic scale if we use a lognormal distribution}

\item{sigma}{the standard deviation of the colony-forming units in the mixed sample on the log10 scale (default value 0.8)}

\item{alpha}{concentration parameter}

\item{k}{number of small portions/ primary samples}

\item{distribution}{what suitable distribution type we have employed for simulation such as \code{"Poisson-Type A"} or \code{"Poisson-Type B"} or \code{"Lognormal-Type A"} or \code{"Lognormal-Type B"}}

\item{n_sim}{number of simulations}

\item{summary}{if we need to get the mean and standard deviation of simulated \eqn{N'}, use summary = TRUE ( default summary =FALSE).}
}
\value{
total number of colony forming units in the single mixing plan
}
\description{
This function calculates the resulting total number of colony forming units in the mixed sample in the single mixing plan. (to be finished later on)
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by mixing of \eqn{k} primary samples and \eqn{N' = \sum N_i}.To more details, please refer the details section of  \link{compare_mixing}. (to be finished later on)
}
\examples{
mu <- 100
sigma <- 0.8
alpha <- 0.1
k <- 10
n_sim <- 20000
sim_single(mu, sigma, alpha, k, distribution = "Poisson-Type A", n_sim, summary = TRUE)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_multiple}, \link{compare_mixing}
}
