% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single.R
\name{sim_single}
\alias{sim_single}
\title{The total number of colony-forming units in the mixed sample by simulation result (in the single mixing plan).}
\usage{
sim_single(n_iter, mu, sigma, b, k, distribution, summary = FALSE)
}
\arguments{
\item{n_iter}{number of iterations}

\item{mu}{average number of colony-forming units in a primary sample which is in logarithmic scale if we use a lognormal distribution}

\item{sigma}{log standard deviation of the colony-forming units in a primary sample}

\item{b}{concentration parameter}

\item{k}{number of small portions/ primary samples}

\item{distribution}{what suitable distribution type we have employed for simulation such as 'Poisson-fair' or 'Poisson-beta' or 'Lognormal-fair' or 'Lognormal-beta'}

\item{summary}{if we need to get the mean and standard deviation of simulated \eqn{N'}, use summary = TRUE ( default summary =FALSE).}
}
\value{
total number of colony forming units in the single mixing plan
}
\description{
This function calculates the resulting total number of colony forming units in the mixed sample in the single mixing plan. (to be finished later on)
}
\details{
Let \eqn{N'} be the number of colony-forming units in the mixed sample which is produced by mixing of \eqn{k} primary sample and \eqn{N' = \sum(N_i)} (to be finished later on)
}
\examples{
n_iter <- 200000
mu <- c(20,30,50,60)
sigma <- c(0.8,0.8,0.8,0.8)
b <- c(0.1,0.1,0.1,0.1)
k <- c(10,10,10,10)
distribution <-  c("Poisson-fair","Poisson-beta","Lognormal-fair","Lognormal-beta")
sim_single( n_iter, mu[2], sigma[2], b[2], k[2], distribution[2], summary = TRUE)
}
\references{
\itemize{
\item Nauta, M.J., 2005. Microbiological risk assessment models for partitioning and mixing during food handling. International Journal of Food Microbiology 100, \href{https://doi.org/10.1016/j.ijfoodmicro.2004.10.027}{311-322}.
}
}
\seealso{
\link{sim_multiple}, \link{compare_mixing}
}
